Combine internal reduce and transduce into _core.  Skip argument checks internally when possible
