+++ core/index.js
@@ -1,7 +1,8 @@
 'use strict'
+var core = require('./_core')
 module.exports = {
-  reduce: require('./reduce'),
-  transduce: require('./transduce'),
+  reduce: core.reduce,
+  transduce: core.transduce,
   eduction: require('./eduction'),
   into: require('./into'),
   sequence: require('./sequence'),
